third_party/blink/web_tests/platform/generic/fast/events/focus-event-source-device-from-keyboard-expected.txt

Test that focus events from keyboard taps should have a sourceCapabilities and its firesTouchEvents should be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

tests on div
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on input
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on button
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on textarea
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on select
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


tests on a
blur
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusout
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focus
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
focusin
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false


PASS successfullyParsed is true

TEST COMPLETE
